requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

langchain==0.3.26
langchain-community==0.3.26
langchain-pinecone==0.2.8
langchain-google-genai==2.1.5

sentence-transformers==4.1.0
python-dotenv==1.1.0

# IMPORTANT: do NOT hard-pin pydantic below LangChain requirement
pydantic>=2.7.4,<3.0.0

pinecone-client==3.0.0

-e .
